services/wave_bridge/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
wandb==0.16.0
anthropic==0.7.8
openai==1.3.5
googletrans==4.0.0rc1
langdetect==1.0.9
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
Jinja2==3.1.2
python-dotenv==1.0.0